Check if user is admin to access backoffice

+++ app/controllers/BackOfficeController.php
@@ -4,25 +4,57 @@ class BackOfficeController extends Controller
 {
     public function index()
     {
-        $this->checkPost();
+        $this->connect();
         $this->redirectIfAdmin();
+    }
+
+    public function connect()
+    {
+        if (!empty($_POST['mail']) || !empty($_POST['password'])) {
+            $user = UserRepository::findByMail($_POST['mail']);
+            if (password_verify($_POST['password'], $user->getPassword())) {
+                $_SESSION['isAdmin'] = true;
+            }
+        }
+    }
+
+    public function redirectIfAdmin()
+    {
+        if (isset($_SESSION['isAdmin'])) {
+            $this->goTobackOfficeView();
+        } else {
+            header("Location: /Admin?error=2");
+            exit();
+        }
     }
 
     public function edit($id)
     {
-        $user = UserRepository::findById($id);
-        $this->view('editUser', ['countries' => CountryRepository::findAllCountriesName(),'user' => $user]);
+        if (isset($_SESSION['isAdmin'])) {
+            $user = UserRepository::findById($id);
+            $this->view('editUser', ['countries' => CountryRepository::findAllCountriesName(), 'user' => $user]);
+        } else {
+            header("Location: /");
+        }
     }
 
     public function delete($id)
     {
-        UserRepository::deleteById($id);
-        $this->goTobackOfficeView();
+        if (isset($_SESSION['isAdmin'])) {
+            UserRepository::deleteById($id);
+            $this->goTobackOfficeView();
+        } else {
+            header("Location: /");
+        }
     }
 
     public function add($countryName)
     {
-        $this->view('index', ['countries' => CountryRepository::findAllCountriesName(), 'countryName' => $countryName]);
+        if (isset($_SESSION['isAdmin'])) {
+            $this->view('index', ['countries' => CountryRepository::findAllCountriesName(), 'countryName' => $countryName]);
+        } else {
+            header("Location: /");
+        }
     }
 
     public function save($id)
@@ -33,40 +65,8 @@ class BackOfficeController extends Controller
         $this->goTobackOfficeView();
     }
 
-    public function checkPost()
-    {
-        if (empty($_POST['mail']) || empty($_POST['password'])) {
-            header("Location: /Admin?error=1");
-            exit();
-        }
-    }
-
-    public function redirectIfAdmin()
-    {
-        $user = UserRepository::findByMail($_POST['mail']);
-        if ($user->isAdmin()) {
-            $this->goTobackOfficeView();
-        } else {
-            header("Location: /Admin?error=2");
-            exit();
-        }
-    }
-
     private function goTobackOfficeView()
     {
-        $this->view('backoffice', ['countriesGroup' => $this->findUsersGroupedByCountry()]);
-    }
-
-    private function findUsersGroupedByCountry()
-    {
-        $countries = UserRepository::findAllCountries();
-        $users = UserRepository::findAllUsers();
-        $array = [];
-        foreach ($countries as $country) {
-            $array[$country] = array_values(array_filter($users, function ($user) use ($country) {
-                return $user['country'] == $country;
-            }));
-        }
-        return $array;
+        $this->view('backoffice', ['countriesGroup' => UserRepository::findUsersGroupedByCountry()]);
     }
 }
